Offer carrier0 as VLAN 7 on eth1 (WAN interface)
Many providers require the use of this VLAN tag in order to access their
services. My current DSL modem is capable of adding the tag on its own
which is why this functionality wasn't implemented in RSDSL. However a
GPON-based fiber infrastructure is going to be built soon. These
networks generally use PPPoE just like DSL, but many of them use
different VLAN tags. In my particular case the ISP is already known and
VLAN 7 is correct for both the current and future providers.